feat(openapi): add oauth with predefined values to scalar ui
